requirements.txt
fastapi
uvicorn[standard]
pymongo
python-jose[cryptography]
pydantic-settings
python-dotenv
pytest
flake8
passlib[bcrypt]
httpx
email-validator
python-multipart
slowapi
